+++ CHANGELOG.md
@@ -15,6 +15,7 @@ developers to maintain and readjust their custom modifications on the main proje
 - Add settings to not display the Add to Google Calendar button in the booking confirmation page (#1763)
 - Add new Google Meet setting and meeting link support for appointments while syncing with Google (#1589)
 - Add Select All / None buttons to multi-checkbox fields
+- Add prompt for client notifications so that they are not sent every time an appointment changes
 
 ### Changed